+++ health/micro-ui/web/core/App.js
@@ -104,11 +104,12 @@ if (initializationError) {
     }
   return (
     <DigitUI
-      stateCode={stateCode}
-      enabledModules={enabledModules}
-      moduleReducers={moduleReducers}
-      defaultLanding="employee"
-    />
+    stateCode={stateCode}
+    enabledModules={enabledModules}
+    moduleReducers={moduleReducers}
+    defaultLanding="employee"
+    allowedUserTypes={["employee"]}
+  />
   );
 }