+++ health/micro-ui/web/micro-ui-internals/packages/modules/health-payments/src/components/attendance_inbox/attendance_inbox.js
@@ -75,6 +75,7 @@ const AttendanceInboxComponent = () => {
                 ? data?.attendanceRegister?.map((item, index) => {
                   return {
                     id: item?.registerNumber,
+                    registerId: item?.id,
                     name: selectedProject?.name,
                     boundary: item?.localityCode,
                     status: item?.attendees == null ? 0 : item?.attendees.length || 0,
@@ -120,7 +121,6 @@ const AttendanceInboxComponent = () => {
   /// Update filter criteria and fetch new data.
   const handleFilterUpdate = (newFilter, isSelectedData) => {
     setFilterCriteria(newFilter);
-    setSelectedStatus(StatusEnum.PENDING_FOR_APPROVAL);
 
     const existingPaymentInbox = Digit.SessionStorage.get("paymentInbox");
 
@@ -154,7 +154,7 @@ const AttendanceInboxComponent = () => {
     Digit.SessionStorage.set("paymentInbox", existingData);
 
     // Trigger the approval action
-    triggerAttendanceSearch(newFilter, StatusEnum.PENDING_FOR_APPROVAL);
+    triggerAttendanceSearch(newFilter, selectedStatus);
   };
 
   const handlePaginationChange = (page) => {
@@ -183,6 +183,7 @@ const AttendanceInboxComponent = () => {
 
   // Reset the table and clear filters.
   const resetTable = () => {
+    setSelectedStatus(StatusEnum.PENDING_FOR_APPROVAL);
     setchildrenData([]);
     setFilterCriteria(null);
     setCard(false);
@@ -193,7 +194,7 @@ const AttendanceInboxComponent = () => {
       <div className="custom-register-inbox-screen">
         <div className="inner-div-row-section">
           <div className="custom-inbox-filter-section">
-            <div className="custom-inbox-inner-filter-section" >
+            <div className="custom-inbox-inner-filter-section" style={{ height: "80vh" }}>
               <CustomFilter
                 resetTable={resetTable}
                 isRequired={ScreenTypeEnum.REGISTER}
@@ -202,8 +203,8 @@ const AttendanceInboxComponent = () => {
             </div>
           </div>
 
-          <div className="custom-inbox-outer-table-section">
-            <div className="inner-table-section">
+          <div className="custom-inbox-outer-table-section" >
+            <div className="inner-table-section" style={{ height: "80vh" }}>
               {card == false ? (
                 <Card className="card-overide">
                   <div className="summary-sub-heading">{t(selectedProject?.name)}</div>